chore: update type definition for Schema Editor (#3842)
