+++ frontend/src/types/schemaEditor/index.ts
@@ -1,40 +1,5 @@
-import { Database, DatabaseId, TableEngineType, TableType } from ".";
-
-type TableOrColumnStatus = "normal" | "created" | "dropped";
-
-export interface Column {
-  // Related fields
-  databaseId: DatabaseId;
-
-  // Domain specific fields
-  oldName: string;
-  newName: string;
-  type: string;
-  nullable: boolean;
-  comment: string;
-  default: string | null;
-
-  status: TableOrColumnStatus;
-}
-
-export interface Table {
-  // Related fields
-  databaseId: DatabaseId;
-
-  // Domain specific fields
-  oldName: string;
-  newName: string;
-  type: TableType;
-  engine: TableEngineType;
-  collation: string;
-  rowCount: number;
-  dataSize: number;
-  comment: string;
-  columnList: Column[];
-  originColumnList: Column[];
-
-  status: TableOrColumnStatus;
-}
+import { Database, DatabaseId } from "..";
+import { Table } from "./atomType";
 
 export enum SchemaEditorTabType {
   TabForDatabase = "database",
