+++ frontend/src/utils/schemaEditor/diffTable.ts
@@ -5,7 +5,7 @@ import type {
   DropTableContext,
   RenameTableContext,
 } from "@/types";
-import { Table } from "@/types/schemaEditor";
+import { Table } from "@/types/schemaEditor/atomType";
 import { transformTableToCreateTableContext } from "./transform";
 import { diffColumnList } from "./diffColumn";
