+++ frontend/src/utils/schemaEditor/transform.ts
@@ -6,7 +6,7 @@ import {
   Column as ColumnData,
   Table as TableData,
 } from "@/types";
-import { Column, Table } from "@/types/schemaEditor";
+import { Column, Table } from "@/types/schemaEditor/atomType";
 
 export const transformTableDataToTable = (tableData: TableData): Table => {
   const columnList = tableData.columnList.map((columnData) =>
